backend/app/requirements.txt
fastapi
argon2-cffi
sqlalchemy
pydantic
uvicorn
python-jose[cryptography]
python-multipart
